introduce MarkdownReader
